+++ instrumentation/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/exporters/logging/LoggingMetricExporterAutoConfiguration.java
@@ -11,15 +11,15 @@ import io.opentelemetry.instrumentation.spring.autoconfigure.exporters.internal.
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Condition;
+import org.springframework.context.annotation.ConditionContext;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.type.AnnotatedTypeMetadata;
 
 /** Configures {@link LoggingMetricExporter} bean for tracing. */
 @Configuration
-@EnableConfigurationProperties(LoggingExporterProperties.class)
 @AutoConfigureBefore(OpenTelemetryAutoConfiguration.class)
 @Conditional(LoggingMetricExporterAutoConfiguration.CustomCondition.class)
 @ConditionalOnClass(LoggingMetricExporter.class)
@@ -33,9 +33,7 @@ public class LoggingMetricExporterAutoConfiguration {
 
   static final class CustomCondition implements Condition {
     @Override
-    public boolean matches(
-        org.springframework.context.annotation.ConditionContext context,
-        org.springframework.core.type.AnnotatedTypeMetadata metadata) {
+    public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
       return ExporterConfigEvaluator.isExporterEnabled(
           context.getEnvironment(),
           "otel.exporter.logging.enabled",