+++ instrumentation/spring/spring-boot-autoconfigure/src/test/java/io/opentelemetry/instrumentation/spring/autoconfigure/exporters/otlp/OtlpMetricExporterAutoConfigurationTest.java
@@ -8,9 +8,7 @@ package io.opentelemetry.instrumentation.spring.autoconfigure.exporters.otlp;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.opentelemetry.exporter.otlp.http.metrics.OtlpHttpMetricExporter;
-import io.opentelemetry.exporter.otlp.metrics.OtlpGrpcMetricExporter;
 import io.opentelemetry.instrumentation.spring.autoconfigure.OpenTelemetryAutoConfiguration;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
@@ -27,28 +25,6 @@ class OtlpMetricExporterAutoConfigurationTest {
   void otlpEnabled() {
     runner
         .withPropertyValues("otel.exporter.otlp.enabled=true")
-        .run(
-            context ->
-                assertThat(context.getBean("otelOtlpMetricExporter", OtlpHttpMetricExporter.class))
-                    .isNotNull());
-  }
-
-  @Test
-  @DisplayName("use grpc when protocol set")
-  void useGrpc() {
-    runner
-        .withPropertyValues("otel.exporter.otlp.protocol=grpc")
-        .run(
-            context ->
-                assertThat(context.getBean("otelOtlpMetricExporter", OtlpGrpcMetricExporter.class))
-                    .isNotNull());
-  }
-
-  @Test
-  @DisplayName("use http when unknown protocol set")
-  void useHttpWhenAnUnknownProtocolIsSet() {
-    runner
-        .withPropertyValues("otel.exporter.otlp.protocol=unknown")
         .run(
             context ->
                 assertThat(context.getBean("otelOtlpMetricExporter", OtlpHttpMetricExporter.class))
@@ -84,13 +60,5 @@ class OtlpMetricExporterAutoConfigurationTest {
     runner
         .withPropertyValues("otel.metrics.exporter=none")
         .run(context -> assertThat(context.containsBean("otelOtlpMetricExporter")).isFalse());
-  }
-
-  @Test
-  void otlpHttpUsedByDefault() {
-    runner.run(
-        context ->
-            assertThat(context.getBean("otelOtlpMetricExporter", OtlpHttpMetricExporter.class))
-                .isNotNull());
   }
 }