+++ instrumentation/spring/spring-boot-autoconfigure/src/test/java/io/opentelemetry/instrumentation/spring/autoconfigure/resources/SpringConfigPropertiesTest.java
@@ -8,13 +8,14 @@ package io.opentelemetry.instrumentation.spring.autoconfigure.resources;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
+import io.opentelemetry.instrumentation.spring.autoconfigure.exporters.otlp.OtlpExporterProperties;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.core.env.Environment;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
 
-class SpringResourceConfigPropertiesTest {
+class SpringConfigPropertiesTest {
   private final ApplicationContextRunner contextRunner = new ApplicationContextRunner();
 
   @Test
@@ -26,8 +27,9 @@ class SpringResourceConfigPropertiesTest {
         .run(
             context -> {
               Environment env = context.getBean("environment", Environment.class);
-              SpringResourceConfigProperties config =
-                  new SpringResourceConfigProperties(env, new SpelExpressionParser());
+              SpringConfigProperties config =
+                  new SpringConfigProperties(
+                      env, new SpelExpressionParser(), new OtlpExporterProperties());
 
               assertThat(config.getMap("otel.springboot.test.map"))
                   .contains(